fix: float deltaTime -> Uint64 deltaTime

+++ src/GameCompleteScene.cpp
@@ -24,7 +24,7 @@ void GameCompleteScene::handleEvent(const SDL_Event& event) {
 	}
 }
 
-void GameCompleteScene::update(float deltaTime) {}
+void GameCompleteScene::update(Uint64 deltaTime) {}
 
 void GameCompleteScene::render(SDL_Renderer* renderer) {
 	TTF_Font* font = m_resourceManager->loadFont("arial.ttf");
