+++ src/LevelScene.h
@@ -10,7 +10,7 @@ class LevelScene : public Scene {
 public:
 	LevelScene(ResourceManager* resourceManager, int levelNum);
 	void handleEvent(const SDL_Event& event) override;
-	void update(float deltaTime) override;
+	void update(Uint64 deltaTime) override;
 	void render(SDL_Renderer* renderer) override;
 private:
 	ResourceManager* m_resourceManager;
